upgrade perf + refractor
